nblast v1 added from Kohl et al 13 - code refactoring
